refactor: mark flag as required function
