+++ cmd/backup/status.go
@@ -67,8 +67,7 @@ func (s *statusCommand) PreRunE(cmd *cobra.Command, _ []string) error {
 
 	if s.clientConfig == nil {
 		s.logger = logger.NewDefaultLogger()
-		cmd.MarkFlagRequired("project-name")
-		cmd.MarkFlagRequired("host")
+		markFlagsRequired(cmd, []string{"project-name", "host"})
 		return nil
 	}
