+++ cmd/namespace/list.go
@@ -68,8 +68,7 @@ func (l *listCommand) PreRunE(cmd *cobra.Command, _ []string) error {
 
 	if l.clientConfig == nil {
 		l.logger = logger.NewDefaultLogger()
-		cmd.MarkFlagRequired("project-name")
-		cmd.MarkFlagRequired("host")
+		markFlagsRequired(cmd, []string{"project-name", "host"})
 		return nil
 	}
