+++ cmd/replay/create.go
@@ -106,8 +106,7 @@ func (c *createCommand) PreRunE(cmd *cobra.Command, _ []string) error {
 	if c.clientConfig == nil {
 		c.logger = logger.NewDefaultLogger()
 		c.survey = survey.NewReplayCreateSurvey(c.logger)
-		cmd.MarkFlagRequired("project-name")
-		cmd.MarkFlagRequired("host")
+		markFlagsRequired(cmd, []string{"project-name", "host"})
 		return nil
 	}
