+++ cmd/replay/replay.go
@@ -28,3 +28,10 @@ func NewReplayCommand() *cobra.Command {
 	)
 	return cmd
 }
+
+// TODO: move it to another common package, eg. internal
+func markFlagsRequired(cmd *cobra.Command, flagNames []string) {
+	for _, n := range flagNames {
+		cmd.MarkFlagRequired(n)
+	}
+}
