+++ cmd/secret/delete.go
@@ -65,8 +65,7 @@ func (d *deleteCommand) PreRunE(cmd *cobra.Command, _ []string) error {
 
 	if d.clientConfig == nil {
 		d.logger = logger.NewDefaultLogger()
-		cmd.MarkFlagRequired("project-name")
-		cmd.MarkFlagRequired("host")
+		markFlagsRequired(cmd, []string{"project-name", "host"})
 		return nil
 	}
