+++ cmd/secret/secret.go
@@ -72,3 +72,10 @@ func validateProperlyEncoded(secretValue string) error {
 	}
 	return nil
 }
+
+// TODO: move it to another common package, eg. internal
+func markFlagsRequired(cmd *cobra.Command, flagNames []string) {
+	for _, n := range flagNames {
+		cmd.MarkFlagRequired(n)
+	}
+}
