+++ cmd/secret/set.go
@@ -83,8 +83,7 @@ func (s *setCommand) PreRunE(cmd *cobra.Command, _ []string) error {
 	if s.clientConfig == nil {
 		s.logger = logger.NewDefaultLogger()
 		s.survey = survey.NewSecretSetSurvey()
-		cmd.MarkFlagRequired("project-name")
-		cmd.MarkFlagRequired("host")
+		markFlagsRequired(cmd, []string{"project-name", "host"})
 		return nil
 	}
